+++ apps/vaxx/modules/2_denominator_selection/survey_comparison.R
@@ -19,8 +19,6 @@ surveyComparisonServer <- function(id, cache, admin_level, region, i18n) {
   moduleServer(
     id = id,
     module = function(input, output, session) {
-      denominatorInputServer("denominator", cache, i18n)
-
       tabPanelsServer(
         "panel",
         serverInput = function(id, current_indicator) {